asset checks revamp
